1.1: Signals and Slots
At this stage, there is a text Label and a QPushButton
When the button is clicked the text: "User Clicked on the Button"
will be displayed in the debug console. The connection when
the button is clicked will modified to add a new task
in the todolist in later commits.